chore(front): fix file directory
