+++ frontend/components/analysis/MetricCard.tsx
