+++ frontend/components/analysis/SOHTrendChart.tsx
