--- frontend/components/dashboard/TrendChart.tsx
@@ -1,3 +1,0 @@
-export default function TrendChart() {
-  return <div></div>;
-}
