--- frontend/components/dashboard/alerts/AlertCard.tsx
@@ -1,5 +1,0 @@
-export default function AlertCard({ message }: { message: string }) {
-  return (
-    <div className='rounded-xl bg-red-900/30 p-4 text-sm'>⚠️ {message}</div>
-  );
-}
